+++ src/main/java/mod/gottsch/neo/claimmyland/core/command/BlockWhitelistSubCommand.java
@@ -78,6 +78,20 @@ public class BlockWhitelistSubCommand extends WhitelistSubCommand {
                                             WhitelistType.BLOCK);
                                 })
                         )
+                )
+                ///// WHITELIST CLEAR /////
+                .then(Commands.literal(CLEAR)
+                        .then(Commands.argument(ESTATE_NAME, StringArgumentType.string())
+                                .suggests(OWNER_ESTATE_NAMES)
+                                .executes(source -> clearFromEstate(source.getSource(),
+                                        StringArgumentType.getString(source, ESTATE_NAME),
+                                        WhitelistType.BLOCK))
+                                .then(Commands.literal(CONFIRM)
+                                        .executes(source -> clearFromEstateConfirmed(source.getSource(),
+                                                StringArgumentType.getString(source, ESTATE_NAME),
+                                                WhitelistType.BLOCK))
+                                )
+                        )
                 );
     }
 
@@ -132,6 +146,25 @@ public class BlockWhitelistSubCommand extends WhitelistSubCommand {
                                                     StringArgumentType.getString(source, ESTATE_NAME),
                                                     WhitelistType.BLOCK);
                                         })
+                                )
+                        )
+                )
+                ///// BLOCK WHITELIST CLEAR /////
+                .then(Commands.literal(CLEAR)
+                        .then(Commands.argument(OWNER_NAME, StringArgumentType.string())
+                                .suggests(OPS_OWNER_NAMES)
+                                .then(Commands.argument(ESTATE_NAME, StringArgumentType.string())
+                                        .suggests(OPS_OWNER_ESTATE_NAMES)
+                                        .executes(source -> clearFromEstate(source.getSource(),
+                                                StringArgumentType.getString(source, OWNER_NAME),
+                                                StringArgumentType.getString(source, ESTATE_NAME),
+                                                WhitelistType.BLOCK))
+                                        .then(Commands.literal(CONFIRM)
+                                                .executes(source -> clearFromEstateConfirmed(source.getSource(),
+                                                        StringArgumentType.getString(source, OWNER_NAME),
+                                                        StringArgumentType.getString(source, ESTATE_NAME),
+                                                        WhitelistType.BLOCK))
+                                        )
                                 )
                         )
                 );