+++ src/main/java/mod/gottsch/neo/claimmyland/core/command/ItemWhitelistSubCommand.java
@@ -77,6 +77,20 @@ public class ItemWhitelistSubCommand extends WhitelistSubCommand {
                                             type);
                                 })
                         )
+                )
+                ///// WHITELIST CLEAR /////
+                .then(Commands.literal(CLEAR)
+                        .then(Commands.argument(ESTATE_NAME, StringArgumentType.string())
+                                .suggests(OWNER_ESTATE_NAMES)
+                                .executes(source -> clearFromEstate(source.getSource(),
+                                        StringArgumentType.getString(source, ESTATE_NAME),
+                                        WhitelistType.ITEM))
+                                .then(Commands.literal(CONFIRM)
+                                        .executes(source -> clearFromEstateConfirmed(source.getSource(),
+                                                StringArgumentType.getString(source, ESTATE_NAME),
+                                                WhitelistType.ITEM))
+                                )
+                        )
                 );
     }
 
@@ -131,6 +145,25 @@ public class ItemWhitelistSubCommand extends WhitelistSubCommand {
                                                     StringArgumentType.getString(source, ESTATE_NAME),
                                                     type);
                                         })
+                                )
+                        )
+                )
+                ///// ITEM WHITELIST CLEAR /////
+                .then(Commands.literal(CLEAR)
+                        .then(Commands.argument(OWNER_NAME, StringArgumentType.string())
+                                .suggests(OPS_OWNER_NAMES)
+                                .then(Commands.argument(ESTATE_NAME, StringArgumentType.string())
+                                        .suggests(OPS_OWNER_ESTATE_NAMES)
+                                        .executes(source -> clearFromEstate(source.getSource(),
+                                                StringArgumentType.getString(source, OWNER_NAME),
+                                                StringArgumentType.getString(source, ESTATE_NAME),
+                                                WhitelistType.ITEM))
+                                        .then(Commands.literal(CONFIRM)
+                                                .executes(source -> clearFromEstateConfirmed(source.getSource(),
+                                                        StringArgumentType.getString(source, OWNER_NAME),
+                                                        StringArgumentType.getString(source, ESTATE_NAME),
+                                                        WhitelistType.ITEM))
+                                        )
                                 )
                         )
                 );